+++ src/components/About/AboutModules.js
@@ -7,9 +7,11 @@ import {
 import css from './About.css';
 
 
-const AboutInterfaces = ({ list }) => {
+const AboutInterfaces = ({ list = [] }) => {
+  list.sort((a, b) => a.name.localeCompare(b.name));
   return (
     <List
+      listStyle="bullets"
       listClass={css.paddingLeftOfListItems}
       items={list}
       itemFormatter={(item) => <li key={item.name}>{item.name}</li>}
@@ -21,7 +23,9 @@ AboutInterfaces.propTypes = {
   list: PropTypes.arrayOf(PropTypes.object),
 };
 
-const AboutModules = ({ list }) => {
+const AboutModules = ({ list = [] }) => {
+  list.sort((a, b) => a.name.localeCompare(b.name));
+
   return (
     <List
       listStyle="bullets"